feat: enhance BaseEnv and RobotEnvFromBase with improved initialization and simulation handling
