+++ robot_sf/gym_env/base_env.py
@@ -7,19 +7,13 @@ Provides common functionality for all robot-based environments.
 # import datetime
 # import pickle
 
-from typing import Callable
+from typing import List
 from loguru import logger
 
 from gymnasium import Env
-from robot_sf.gym_env.reward import simple_reward
 from robot_sf.gym_env.env_config import EnvSettings
-
-# from robot_sf.gym_env.env_config import BaseEnvSettings
-# from robot_sf.render.sim_view import (
-#     SimulationView,
-#     VisualizableSimState,
-# )
-# from robot_sf.sim.simulator import init_simulators
+from robot_sf.render.sim_view import SimulationView, VisualizableSimState
+from robot_sf.sim.simulator import init_simulators
 
 
 class BaseEnv(Env):
@@ -28,12 +22,12 @@ class BaseEnv(Env):
     def __init__(
         self,
         env_config: EnvSettings = EnvSettings(),
-        reward_func: Callable[[dict], float] = simple_reward,
         debug: bool = False,
         recording_enabled: bool = False,
         record_video: bool = False,
         video_path: str = None,
         video_fps: float = None,
+        peds_have_obstacle_forces: bool = False,
     ):
         super().__init__()
 
@@ -48,8 +42,46 @@ class BaseEnv(Env):
         # Extract first map definition; currently only supports using the first map
         self.map_def = env_config.map_pool.choose_random_map()
 
+        self.debug = debug
+
+        # Initialize the list to store recorded states
+        self.recorded_states: List[VisualizableSimState] = []
+        self.recording_enabled = recording_enabled
+
+        # Initialize simulator with a random start position
+        self.simulator = init_simulators(
+            env_config,
+            self.map_def,
+            random_start_pos=True,
+            peds_have_obstacle_forces=peds_have_obstacle_forces,
+        )[0]
+
+        # Store last action executed by the robot
+        self.last_action = None
+
+        # If in debug mode or video recording is enabled, create simulation view
+        if debug or record_video:
+            self.sim_ui = SimulationView(
+                scaling=10,
+                map_def=self.map_def,
+                obstacles=self.map_def.obstacles,
+                robot_radius=env_config.robot_config.radius,
+                ped_radius=env_config.sim_config.ped_radius,
+                goal_radius=env_config.sim_config.goal_radius,
+                record_video=record_video,
+                video_path=video_path,
+                video_fps=video_fps,
+            )
+
     def render(self):
         raise NotImplementedError
 
     def step(self, action):
         raise NotImplementedError
+
+    def exit(self):
+        """
+        Clean up and exit the simulation UI, if it exists.
+        """
+        if self.sim_ui:
+            self.sim_ui.exit_simulation()
